--- lib/generators/rolemodel/optics/base/templates/app/helpers/application_helper.rb
@@ -1,12 +1,0 @@
-module ApplicationHelper
-  def icon_name_for_flash(type)
-    case type
-    when 'notice'
-      'check_circle'
-    when 'alert'
-      'cancel'
-    else
-        type
-    end
-  end
-end
