+++ lib/generators/rolemodel/slim/slim_generator.rb
@@ -3,5 +3,10 @@ module Rolemodel
     def add_slim
       run 'bundle add slim'
     end
+
+    def replace_erb_layout
+      remove_file 'app/views/layouts/application.html.erb'
+      copy_file 'app/views/layouts/application.html.slim'
+    end
   end
 end
