+++ lib/generators/rolemodel/slim/templates/app/views/layouts/application.html.slim
@@ -8,6 +8,5 @@ html
     = javascript_include_tag 'application', 'data-turbo-track': 'reload'
     meta content="width=device-width, initial-scale=1" name="viewport"
   body.app-body
-    = render 'shared/flash'
     .app__content
       = yield
